Unify wrappers into ChainWorkerGuard.
