+++ linera-core/src/chain_worker/delivery_notifier.rs
@@ -22,7 +22,7 @@ use tracing::warn;
 ///
 /// The notifier instance can be cheaply `clone`d and works as a shared reference.
 /// However, its methods still require `&mut self` to hint that it should only be changed by
-/// [`ChainWorkerStateWithAttemptedChanges`](super::ChainWorkerStateWithAttemptedChanges).
+/// [`ChainWorkerGuard`](super::ChainWorkerGuard).
 #[derive(Clone, Default)]
 pub struct DeliveryNotifier {
     notifiers: Arc<Mutex<BTreeMap<BlockHeight, Vec<oneshot::Sender<()>>>>>,
