+++ linera-core/src/chain_worker/state/guard.rs
@@ -8,22 +8,25 @@ use std::{borrow::Cow, collections::BTreeMap};
 use futures::future::Either;
 use linera_base::{
     crypto::ValidatorPublicKey,
-    data_types::{Blob, BlockHeight, Epoch, Round, Timestamp},
+    data_types::{
+        ApplicationDescription, ArithmeticError, Blob, BlockHeight, Epoch, Round, Timestamp,
+    },
     ensure,
-    identifiers::{ChainId, EventId, StreamId},
+    identifiers::{AccountOwner, ApplicationId, ChainId, EventId, StreamId},
 };
 use linera_chain::{
     data_types::{
-        BlockExecutionOutcome, BlockProposal, MessageBundle, OriginalProposal, ProposalContent,
+        BlockExecutionOutcome, BlockProposal, IncomingBundle, MessageAction, MessageBundle,
+        OriginalProposal, ProposalContent, ProposedBlock,
     },
     manager,
-    types::{ConfirmedBlockCertificate, TimeoutCertificate, ValidatedBlockCertificate},
+    types::{Block, ConfirmedBlockCertificate, TimeoutCertificate, ValidatedBlockCertificate},
     ChainExecutionContext, ChainStateView, ExecutionResultExt as _,
 };
-use linera_execution::{committee::Committee, system::EPOCH_STREAM_NAME};
+use linera_execution::{committee::Committee, system::EPOCH_STREAM_NAME, Query, QueryOutcome};
 use linera_storage::{Clock as _, Storage};
 use linera_views::{
-    context::Context,
+    context::{Context, InactiveContext},
     views::{ReplaceContext, RootView, View},
 };
 use tokio::sync::oneshot;
@@ -31,31 +34,27 @@ use tracing::{debug, instrument, trace, warn};
 
 use super::{check_block_epoch, ChainWorkerConfig, ChainWorkerState};
 use crate::{
-    data_types::ChainInfoResponse,
+    data_types::{ChainInfo, ChainInfoQuery, ChainInfoResponse},
     worker::{NetworkActions, Notification, Reason, WorkerError},
 };
 
-/// Wrapper type that tracks if the changes to the `chain` state should be rolled back when
-/// dropped.
-pub struct ChainWorkerStateWithAttemptedChanges<'state, StorageClient>(
-    &'state mut ChainWorkerState<StorageClient>,
-)
+/// Wrapper type that rolls back unsaved changes to the `chain` state when dropped.
+pub struct ChainWorkerGuard<'state, StorageClient>(&'state mut ChainWorkerState<StorageClient>)
 where
     StorageClient: Storage + Clone + Send + Sync + 'static;
 
-impl<'state, StorageClient> ChainWorkerStateWithAttemptedChanges<'state, StorageClient>
+impl<'state, StorageClient> ChainWorkerGuard<'state, StorageClient>
 where
     StorageClient: Storage + Clone + Send + Sync + 'static,
 {
-    /// Creates a new [`ChainWorkerStateWithAttemptedChanges`] instance to change the
-    /// `state`.
+    /// Creates a new [`ChainWorkerGuard`] instance to change the `state`.
     pub(super) async fn new(state: &'state mut ChainWorkerState<StorageClient>) -> Self {
         assert!(
             !state.chain.has_pending_changes().await,
             "`ChainStateView` has unexpected leftover changes"
         );
 
-        ChainWorkerStateWithAttemptedChanges(state)
+        ChainWorkerGuard(state)
     }
 
     /// Processes a leader timeout issued for this multi-owner chain.
@@ -693,6 +692,298 @@ where
         ))
     }
 
+    /// Returns a stored [`Certificate`] for the chain's block at the requested [`BlockHeight`].
+    #[cfg(with_testing)]
+    pub(super) async fn read_certificate(
+        &mut self,
+        height: BlockHeight,
+    ) -> Result<Option<ConfirmedBlockCertificate>, WorkerError> {
+        self.0.ensure_is_active().await?;
+        let certificate_hash = match self.0.chain.confirmed_log.get(height.try_into()?).await? {
+            Some(hash) => hash,
+            None => return Ok(None),
+        };
+        let certificate = self
+            .0
+            .storage
+            .read_certificate(certificate_hash)
+            .await?
+            .ok_or_else(|| WorkerError::ReadCertificatesError(vec![certificate_hash]))?;
+        Ok(Some(certificate))
+    }
+
+    /// Queries an application's state on the chain.
+    pub(super) async fn query_application(
+        &mut self,
+        query: Query,
+    ) -> Result<QueryOutcome, WorkerError> {
+        self.0.ensure_is_active().await?;
+        let local_time = self.0.storage.clock().current_time();
+        let outcome = self
+            .0
+            .chain
+            .query_application(local_time, query, self.0.service_runtime_endpoint.as_mut())
+            .await?;
+        Ok(outcome)
+    }
+
+    /// Returns an application's description.
+    pub(super) async fn describe_application(
+        &mut self,
+        application_id: ApplicationId,
+    ) -> Result<ApplicationDescription, WorkerError> {
+        self.0.ensure_is_active().await?;
+        let response = self.0.chain.describe_application(application_id).await?;
+        Ok(response)
+    }
+
+    /// Executes a block without persisting any changes to the state.
+    pub(super) async fn stage_block_execution(
+        &mut self,
+        block: ProposedBlock,
+        round: Option<u32>,
+        published_blobs: &[Blob],
+    ) -> Result<(Block, ChainInfoResponse), WorkerError> {
+        self.0.ensure_is_active().await?;
+        let local_time = self.0.storage.clock().current_time();
+        let signer = block.authenticated_signer;
+        let (_, committee) = self.0.chain.current_committee()?;
+        block.check_proposal_size(committee.policy().maximum_block_proposal_size)?;
+
+        let outcome = self
+            .execute_block(&block, local_time, round, published_blobs)
+            .await?;
+
+        let mut response = ChainInfoResponse::new(&self.0.chain, None);
+        if let Some(signer) = signer {
+            response.info.requested_owner_balance = self
+                .0
+                .chain
+                .execution_state
+                .system
+                .balances
+                .get(&signer)
+                .await?;
+        }
+
+        Ok((outcome.with(block), response))
+    }
+
+    /// Validates a proposal's signatures; returns `manager::Outcome::Skip` if we already voted
+    /// for it.
+    pub(super) async fn check_proposed_block(
+        &self,
+        proposal: &BlockProposal,
+    ) -> Result<manager::Outcome, WorkerError> {
+        proposal
+            .check_invariants()
+            .map_err(|msg| WorkerError::InvalidBlockProposal(msg.to_string()))?;
+        proposal.check_signature()?;
+        let owner = proposal.owner();
+        let BlockProposal {
+            content,
+            original_proposal,
+            signature: _,
+        } = proposal;
+        let block = &content.block;
+        let chain = &self.0.chain;
+        // Check the epoch.
+        let (epoch, committee) = chain.current_committee()?;
+        super::check_block_epoch(epoch, block.chain_id, block.epoch)?;
+        let policy = committee.policy().clone();
+        block.check_proposal_size(policy.maximum_block_proposal_size)?;
+        // Check the authentication of the block.
+        ensure!(
+            chain.manager.verify_owner(&owner, proposal.content.round)?,
+            WorkerError::InvalidOwner
+        );
+        match original_proposal {
+            None => {
+                if let Some(signer) = block.authenticated_signer {
+                    // Check the authentication of the operations in the new block.
+                    ensure!(signer == owner, WorkerError::InvalidSigner(owner));
+                }
+            }
+            Some(OriginalProposal::Regular { certificate }) => {
+                // Verify that this block has been validated by a quorum before.
+                certificate.check(committee)?;
+            }
+            Some(OriginalProposal::Fast(signature)) => {
+                let original_proposal = BlockProposal {
+                    content: ProposalContent {
+                        block: proposal.content.block.clone(),
+                        round: Round::Fast,
+                        outcome: None,
+                    },
+                    signature: *signature,
+                    original_proposal: None,
+                };
+                let super_owner = original_proposal.owner();
+                ensure!(
+                    chain
+                        .manager
+                        .ownership
+                        .get()
+                        .super_owners
+                        .contains(&super_owner),
+                    WorkerError::InvalidOwner
+                );
+                if let Some(signer) = block.authenticated_signer {
+                    // Check the authentication of the operations in the new block.
+                    ensure!(signer == super_owner, WorkerError::InvalidSigner(signer));
+                }
+                original_proposal.check_signature()?;
+            }
+        }
+        // Check if the chain is ready for this new block proposal.
+        chain.tip_state.get().verify_block_chaining(block)?;
+        Ok(chain.manager.check_proposed_block(proposal)?)
+    }
+
+    /// Validates and executes a block proposed to extend this chain.
+    pub(super) async fn validate_proposal_content(
+        &mut self,
+        content: &ProposalContent,
+        published_blobs: &[Blob],
+    ) -> Result<(BlockExecutionOutcome, Timestamp), WorkerError> {
+        let ProposalContent {
+            block,
+            round,
+            outcome,
+        } = content;
+
+        let local_time = self.0.storage.clock().current_time();
+        ensure!(
+            block.timestamp.duration_since(local_time) <= self.0.config.grace_period,
+            WorkerError::InvalidTimestamp
+        );
+        self.0.storage.clock().sleep_until(block.timestamp).await;
+        let local_time = self.0.storage.clock().current_time();
+
+        self.0
+            .chain
+            .remove_bundles_from_inboxes(block.timestamp, block.incoming_bundles())
+            .await?;
+        let outcome = if let Some(outcome) = outcome {
+            outcome.clone()
+        } else {
+            self.execute_block(block, local_time, round.multi_leader(), published_blobs)
+                .await?
+        };
+
+        ensure!(
+            !round.is_fast() || !outcome.has_oracle_responses(),
+            WorkerError::FastBlockUsingOracles
+        );
+        let chain = &mut self.0.chain;
+        // Check if the counters of tip_state would be valid.
+        chain
+            .tip_state
+            .get_mut()
+            .update_counters(&block.transactions, &outcome.messages)?;
+        // Verify that the resulting chain would have no unconfirmed incoming messages.
+        chain.validate_incoming_bundles().await?;
+        Ok((outcome, local_time))
+    }
+
+    /// Prepares a [`ChainInfoResponse`] for a [`ChainInfoQuery`].
+    pub(super) async fn prepare_chain_info_response(
+        &mut self,
+        query: ChainInfoQuery,
+    ) -> Result<ChainInfoResponse, WorkerError> {
+        self.0.ensure_is_active().await?;
+        let chain = &self.0.chain;
+        let mut info = ChainInfo::from(chain);
+        if query.request_committees {
+            info.requested_committees = Some(chain.execution_state.system.committees.get().clone());
+        }
+        if query.request_owner_balance == AccountOwner::CHAIN {
+            info.requested_owner_balance = Some(*chain.execution_state.system.balance.get());
+        } else {
+            info.requested_owner_balance = chain
+                .execution_state
+                .system
+                .balances
+                .get(&query.request_owner_balance)
+                .await?;
+        }
+        if let Some(next_block_height) = query.test_next_block_height {
+            ensure!(
+                chain.tip_state.get().next_block_height == next_block_height,
+                WorkerError::UnexpectedBlockHeight {
+                    expected_block_height: next_block_height,
+                    found_block_height: chain.tip_state.get().next_block_height
+                }
+            );
+        }
+        if query.request_pending_message_bundles {
+            let mut messages = Vec::new();
+            let pairs = chain.inboxes.try_load_all_entries().await?;
+            let action = if *chain.execution_state.system.closed.get() {
+                MessageAction::Reject
+            } else {
+                MessageAction::Accept
+            };
+            for (origin, inbox) in pairs {
+                for bundle in inbox.added_bundles.elements().await? {
+                    messages.push(IncomingBundle {
+                        origin,
+                        bundle,
+                        action,
+                    });
+                }
+            }
+
+            info.requested_pending_message_bundles = messages;
+        }
+        if let Some(range) = query.request_sent_certificate_hashes_in_range {
+            let start: usize = range.start.try_into()?;
+            let end = match range.limit {
+                None => chain.confirmed_log.count(),
+                Some(limit) => start
+                    .checked_add(usize::try_from(limit).map_err(|_| ArithmeticError::Overflow)?)
+                    .ok_or(ArithmeticError::Overflow)?
+                    .min(chain.confirmed_log.count()),
+            };
+            let keys = chain.confirmed_log.read(start..end).await?;
+            info.requested_sent_certificate_hashes = keys;
+        }
+        if let Some(start) = query.request_received_log_excluding_first_n {
+            let start = usize::try_from(start).map_err(|_| ArithmeticError::Overflow)?;
+            info.requested_received_log = chain.received_log.read(start..).await?;
+        }
+        if query.request_manager_values {
+            info.manager.add_values(&chain.manager);
+        }
+        Ok(ChainInfoResponse::new(info, self.0.config.key_pair()))
+    }
+
+    /// Executes a block, caches the result, and returns the outcome.
+    async fn execute_block(
+        &mut self,
+        block: &ProposedBlock,
+        local_time: Timestamp,
+        round: Option<u32>,
+        published_blobs: &[Blob],
+    ) -> Result<BlockExecutionOutcome, WorkerError> {
+        let outcome =
+            Box::pin(
+                self.0
+                    .chain
+                    .execute_block(block, local_time, round, published_blobs, None),
+            )
+            .await?;
+        self.0.execution_state_cache.insert_owned(
+            &outcome.state_hash,
+            self.0
+                .chain
+                .execution_state
+                .with_context(|ctx| InactiveContext(ctx.base_key().clone()))
+                .await,
+        );
+        Ok(outcome)
+    }
+
     /// Stores the chain state in persistent storage.
     ///
     /// Waits until the [`ChainStateView`] is no longer shared before persisting the changes.
@@ -703,7 +994,7 @@ where
     }
 }
 
-impl<StorageClient> Drop for ChainWorkerStateWithAttemptedChanges<'_, StorageClient>
+impl<StorageClient> Drop for ChainWorkerGuard<'_, StorageClient>
 where
     StorageClient: Storage + Clone + Send + Sync + 'static,
 {
